+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.8.6
+* CXV1-21338 - Put disconnect check back on queue if timeout hasn't ended
+* CXV1-21338 - Fix conversation events timestamp on client activity.
+
 ## 1.8.5
 * CXV1-21499 - Put back agent disconnect message in customer disconnect.